requirements.txt
a2wsgi==1.10.8
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.1
apache-airflow-core==3.0.1
apache-airflow-providers-celery==3.10.6
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.27.0
apache-airflow-providers-fab==2.0.2
apache-airflow-providers-postgres==6.1.3
apache-airflow-providers-sftp==5.2.1
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-ssh==4.0.1
apache-airflow-providers-standard==1.1.0
apache-airflow-task-sdk==1.0.1
apispec==6.8.1
argcomplete==3.6.2
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
Automat==22.10.0
babel==2.17.0
bcc==0.29.1
bcrypt==4.3.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
boto3==1.34.46
botocore==1.34.46
cachelib==0.13.0
cadwyn==5.3.3
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloud-init==24.4.1
colorama==0.4.6
colorlog==6.9.0
command-not-found==0.3
configobj==5.0.8
connexion==2.14.2
constantly==23.10.4
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
dbus-python==1.3.2
Deprecated==1.2.18
dill==0.3.1.1
distro==1.9.0
distro-info==1.7+build1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flower==2.0.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httptools==0.6.4
httpx==0.27.0
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.4.0
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lazy-object-proxy==1.11.0
libcst==1.7.0
limits==5.1.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==0.28.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==25.0
pandas==2.1.4
paramiko==3.5.1
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==4.25.7
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.1.0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
retryhttp==1.3.3
rich==13.9.4
rich-argparse==1.7.0
rich-toolkit==0.14.5
rpds-py==0.24.0
s3transfer==0.10.1
service-identity==24.1.0
setproctitle==1.3.6
setuptools==68.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sos==4.8.2
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
ssh-import-id==5.11
sshtunnel==0.4.0
starlette==0.46.2
structlog==25.3.0
svcs==25.1.0
systemd-python==235
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tornado==6.4.2
Twisted==24.3.0
typer==0.15.3
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.2
unattended-upgrades==0.1
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
wadllib==1.3.6
watchfiles==1.0.5
wcwidth==0.2.13
websockets==14.2
Werkzeug==2.2.3
wheel==0.42.0
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
xkit==0.0.0
zipp==3.21.0
zope.interface==6.1
beautifulsoup4~=4.12.3
zeep~=4.2.1
retry~=0.9.2
